Introducing control.signals package
