+++ node/testkit/src/test/scala/fluence/node/BlockUploadingIntegrationSpec.scala
@@ -52,7 +52,7 @@ import fluence.node.workers.tendermint.block.BlockUploading
 import fluence.node.workers.tendermint.config.{ConfigTemplate, TendermintConfig}
 import fluence.node.workers.{Worker, WorkerBlockManifests, WorkerParams, WorkerServices}
 import fluence.statemachine.AbciService.TxResponse
-import fluence.statemachine.control.{BlockReceipt, ControlSignals}
+import fluence.statemachine.control.signals.{BlockReceipt, ControlSignals}
 import fluence.statemachine.data.{Tx, TxCode}
 import fluence.statemachine.error.StateMachineError
 import fluence.statemachine.state.AbciState
