+++ statemachine/src/main/scala/fluence/statemachine/ServerRunner.scala
@@ -31,12 +31,13 @@ import fluence.effects.tendermint.rpc.TendermintRpc
 import fluence.effects.tendermint.rpc.http.TendermintHttpRpc
 import fluence.log.{Log, LogFactory, LogLevel}
 import fluence.statemachine.config.StateMachineConfig
-import fluence.statemachine.control.{ControlServer, ControlSignals}
+import fluence.statemachine.control.ControlServer
 import fluence.statemachine.error.StateMachineError
 import fluence.statemachine.vm.WasmVmOperationInvoker
 import fluence.vm.WasmVm
 import fluence.vm.wasm.MemoryHasher
 import LogLevel.toLogLevel
+import fluence.statemachine.control.signals.ControlSignals
 
 import scala.language.higherKinds
