+++ statemachine/src/test/scala/fluence/statemachine/StatemachineIntegrationSpec.scala
@@ -28,7 +28,7 @@ import fluence.effects.tendermint.rpc.http.{TendermintHttpRpc, TendermintHttpRpc
 import fluence.log.{Log, LogFactory}
 import fluence.statemachine.config.{StateMachineConfig, TendermintRpcConfig}
 import fluence.statemachine.control.ControlServer
-import fluence.statemachine.control.{ControlSignals, MockedControlSignals}
+import fluence.statemachine.control.signals.{ControlSignals, MockedControlSignals}
 import fluence.statemachine.data.{QueryCode, TxCode}
 import org.scalatest.{Matchers, OneInstancePerTest, WordSpec}
 import scodec.bits.ByteVector
